+++ Assets/Project Designer+/Editor/Core/ProjectDesignerPackageInfo.cs
@@ -10,7 +10,7 @@ namespace ProjectDesigner.V2.Editor
 
         public static string GetInstalledVersion()
         {
-            return "3.0.0";
+            return "3.1.0";
         }
     }
 }